requirements.txt
fastapi 
uvicorn 
smolagents[litellm] 
python-multipart
gradio_client
mistralai
Pillow
requests
python-dotenv